onboarding script now fetches profile and inserts it into postgres db. Add README instructions.
