+++ onboarding/requirements.txt
@@ -1,0 +1,4 @@
+databases[postgresql]==0.4.2
+SQLAlchemy==1.3.16
+psycopg2
+alembic==1.4.2
